Better handling of input in `Output#stream`.
While hijacking responses was working okay, hijacking requests was broken.
`@stream.input` was not set when invoking `send_body`, so the input stream
was never available. We wait for the response to call `notify!` which
ensures the response input is available. To avoid a race condition, we
use `@response.body` rather than `@input` which could be null if an
`end_stream` frame was already processed.