Fix off-by-1 in RelativeLocation.
The number of redirects allowed by the RelativeLocation class was off by
one. A hop is a redirect, but if you passed in 0 hops it wasn't allowing
even a single request. It should have allowed one request but no
redirects.

Also changed the default maximum_hops from 4 to 3 to keep the default
behavior consistent with what it was before.