Endpoint.for should normalize URL
When using Endpoint.parse, the generated URL is normalized before
passing to the constructor, resulting in a `path` attribute of `/` by default.

When using Endpoint.for, the path is left as an empty String, which causes
issues in some web servers. To make the behaviour consistent, Endpoint.for
will normalize as well.